Add `ActionError` impls for `AuthError`
